requirements.txt
Django==6.0
djangorestframework==3.16.1
drf-yasg==1.21.15
psycopg2-binary==2.9.11
django-cors-headers==4.9.0
python-dotenv==1.2.1
django-filter==25.1
django-jazzmin==3.0.1
djangorestframework-simplejwt==5.3.1
django-allauth==0.63.3
gunicorn==23.0.0
redis==5.2.1
django-redis==5.4.0
